* Evento.cs: * Registro.cs: * Trabajador.cs: * TipoRegistro.cs: * Recomendacion.cs: * PerfilUsuario.cs: * CreditoTipoCasa.cs: * Stprm.DataEx.csproj:

+++ branches/Stprm/WebApp/Stprm.DataEx/CreditoTipoCasa.cs
@@ -11,7 +11,7 @@ namespace Stprm.DataEx
 
 		public int Id;
 		public string TipoCasa;
-		public decimal MontoCredito;
+		public long MontoCredito;
 		public string Descripcion;
 		
 		public CreditoTipoCasa (BaseDatos datos) : base (datos, TipoRegistro.TipoCasa)
